Fresh Textfield with singleline will cause a crash if gettextbycoordinate is called thus we set maxlinecount and preferedlinecount to 1 and avoid it being 0

+++ Export_LemurChanges/src/TextField.java
@@ -239,7 +239,7 @@ public class TextField extends Panel {
     }
 
 
-    public float getPreferredLineCount() {
+    public int getPreferredLineCount() {
         return text.getPreferredLineCount();
     }
